# gpt/actions_openapi.yaml
openapi: 3.1.0
info:
  title: Flooring Hotspot GPT — Middleware
  description: |
    Middleware API called by the Flooring Hotspot ChatGPT GPT via Actions.
    Wraps Semrush, normalizes/clusters/scores topics for the SPC / PVC /
    LVT / Vinyl / Commercial Flooring B2B export use case.
  version: "0.1.0"

servers:
  - url: https://your-public-host.example.com
    description: Replace with your deployed PUBLIC_BASE_URL (must be HTTPS for GPT Actions).

security:
  - bearerAuth: []

paths:
  /analyze:
    post:
      operationId: analyzeHotTopics
      summary: Analyze competitor flooring hot topics via Semrush
      description: |
        Given competitor domains, calls Semrush (or mock fallback), clusters
        the returned pages and keywords into flooring-industry topics, and
        returns a ranked list of hot topics with explainable scores plus
        suggested Blog / LinkedIn / cold-email angles.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRequest"
      responses:
        "200":
          description: Ranked list of hot topics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeResponse"
        "400":
          description: Bad request
        "401":
          description: Missing bearer token
        "403":
          description: Invalid bearer token

  /export:
    post:
      operationId: exportTopics
      summary: Export a topic list to CSV / XLSX / JSON
      description: |
        Pass the `topics` array returned by /analyze (or /import-csv) and
        the desired format. Returns a `download_url` valid for ~24h.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportRequest"
      responses:
        "200":
          description: Download URL of the generated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResponse"

  /health:
    get:
      operationId: getHealth
      summary: Liveness probe
      security: []   # no auth required
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  data_source_mode: { type: string }
                  semrush_configured: { type: boolean }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    AnalyzeRequest:
      type: object
      required: [competitor_domains]
      properties:
        user_token:
          type: string
          nullable: true
          description: BYOK access code from /setup, format `floor-XXXX-XXXX`. Required in BYOK mode.
          example: "floor-7K3Q-9WX2"
        competitor_domains:
          type: array
          minItems: 1
          maxItems: 10
          items: { type: string }
          description: Competitor root domains, e.g. ["shaw.com", "mohawkflooring.com"]
          example: ["shaw.com", "mohawkflooring.com"]
        product_focus:
          type: string
          enum: [spc, pvc, lvt, vinyl, commercial, any]
          default: any
        country:
          type: string
          minLength: 2
          maxLength: 2
          default: us
          description: Two-letter Semrush database code.
        time_window_days:
          type: integer
          enum: [30, 60, 90]
          default: 90
        output_use_case:
          type: string
          enum: [blog, linkedin, cold_email, general]
          default: general
        top_n:
          type: integer
          minimum: 5
          maximum: 50
          default: 20

    TopicSupportPage:
      type: object
      properties:
        competitor_domain: { type: string }
        page_url: { type: string }
        page_title: { type: string, nullable: true }
        estimated_traffic: { type: number }

    TopicResult:
      type: object
      properties:
        canonical_topic: { type: string }
        family: { type: string }
        competitor_count: { type: integer }
        page_count: { type: integer }
        keyword_count: { type: integer }
        related_keywords:
          type: array
          items: { type: string }
        supporting_pages:
          type: array
          items: { $ref: "#/components/schemas/TopicSupportPage" }
        freshness_window_days: { type: integer }
        hotness_score: { type: integer, minimum: 0, maximum: 100 }
        buyer_relevance_score: { type: integer, minimum: 0, maximum: 100 }
        product_fit_score: { type: integer, minimum: 0, maximum: 100 }
        opportunity_score: { type: integer, minimum: 0, maximum: 100 }
        score_explanation: { type: string }
        suggested_blog_angle: { type: string }
        suggested_linkedin_angle: { type: string }
        suggested_cold_email_angle: { type: string }
        why_it_matters: { type: string }

    AnalyzeResponse:
      type: object
      properties:
        data_source:
          type: string
          enum: [semrush, mock, csv]
          description: GPT must surface this to the user when not "semrush".
        competitor_domains:
          type: array
          items: { type: string }
        country: { type: string }
        time_window_days: { type: integer }
        product_focus: { type: string }
        output_use_case: { type: string }
        total_pages_analyzed: { type: integer }
        total_keywords_analyzed: { type: integer }
        topics:
          type: array
          items: { $ref: "#/components/schemas/TopicResult" }
        notes:
          type: array
          items: { type: string }
        quota_used_today:
          type: integer
          nullable: true
          description: Calls used today against this user_token. Null in non-BYOK modes.
        quota_limit_today:
          type: integer
          nullable: true

    ExportRequest:
      type: object
      required: [topics]
      properties:
        topics:
          type: array
          items: { $ref: "#/components/schemas/TopicResult" }
        format:
          type: string
          enum: [csv, xlsx, json]
          default: xlsx
        filename_hint:
          type: string
          nullable: true

    ExportResponse:
      type: object
      properties:
        download_url: { type: string }
        format: { type: string }
        rows: { type: integer }